Доделал функции контроллера.

+++ TP_Lab2/Models/Person.cs
@@ -6,6 +6,6 @@
         public string Name { get; set; }
         public string Login { get; set; }
         public string Password { get; set; }
-        public int PurchasesSum { get; set; }
+        public float PurchasesSum { get; set; }
     }
 }
